The Lamb Lies Down On Broadway intro finished
The Lamb Lies Down On Broadway is a test of if the music core can run quickly and evenly across two separate robots at the same time. Also it is a cool piano part. Also the speaker can't properly play notes below 300 Hz so a few of the notes at the end of the second layer are replaced with rests.